requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-jose==3.3.0
python-dotenv==1.0.1
slowapi==0.1.9

